backend/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
importlib-metadata==4.0.1
joblib==1.0.1
numpy==1.20.2
packaging==20.9
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
six==1.15.0
sqlparse==0.4.1
tokenizers==0.10.2
torch==1.8.1+cpu
tqdm==4.60.0
transformers==4.5.1
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
